+++ src/wally/manifest.ts
@@ -4,6 +4,8 @@ import { matchDependencyPartial } from "../utils/matching";
 
 import { parseTomlTokens } from "../utils/toml";
 
+import { WallyPackageRealm } from "./base";
+
 
 
 
@@ -200,4 +202,27 @@ export const parseWallyManifest = (document: vscode.TextDocument) => {
 		}
 	}
 	return manifest;
+};
+
+export const findDependencyAtPosition = (
+	manifest: WallyManifest,
+	position: vscode.Position
+): ([
+	WallyPackageRealm,
+	WallyManifestDependency
+] | null) => {
+	const allDeps: [WallyPackageRealm, WallyManifestDependency[]][] = [
+		["shared", manifest.dependencies.shared],
+		["server", manifest.dependencies.server],
+		["dev", manifest.dependencies.dev],
+	];
+	for (const [dependencyRealm, dependencyList] of allDeps) {
+		for (const dependency of dependencyList) {
+			const deprange = new vscode.Range(dependency.start, dependency.end);
+			if (deprange.contains(position)) {
+				return [dependencyRealm, dependency];
+			}
+		}
+	}
+	return null;
 };